+++ examples/iMAML/imaml_omniglot_functional.py
@@ -24,6 +24,7 @@ for few-shot Omniglot classification.
 import argparse
 import time
 
+import functorch
 import matplotlib as mpl
 import matplotlib.pyplot as plt
 import numpy as np
@@ -33,7 +34,7 @@ import torch.nn as nn
 import torch.nn.functional as F
 
 import torchopt
-from torchopt.diff.implicit import ImplicitMetaGradientModule
+from torchopt import pytree
 
 
 from helpers.omniglot_loaders import OmniglotNShot  # isort: skip
@@ -41,41 +42,6 @@ from helpers.omniglot_loaders import OmniglotNShot  # isort: skip
 
 mpl.use('Agg')
 plt.style.use('bmh')
-
-
-class InnerNet(
-    ImplicitMetaGradientModule,
-    linear_solve=torchopt.linear_solve.solve_normal_cg(maxiter=5, atol=0),
-):
-    def __init__(self, meta_net, n_inner_iter, reg_param):
-        super().__init__()
-        self.meta_net = meta_net
-        self.net = torchopt.module_clone(meta_net, by='deepcopy', detach_buffers=True)
-        self.n_inner_iter = n_inner_iter
-        self.reg_param = reg_param
-
-    def forward(self, x):
-        return self.net(x)
-
-    def objective(self, x, y):
-        y_pred = self(x)
-        loss = F.cross_entropy(y_pred, y)
-        regularization_loss = 0
-        for p1, p2 in zip(self.parameters(), self.meta_parameters()):
-            regularization_loss += 0.5 * self.reg_param * torch.sum(torch.square(p1 - p2))
-        return loss + regularization_loss
-
-    def solve(self, x, y):
-        params = tuple(self.parameters())
-        inner_optim = torchopt.SGD(params, lr=1e-1)
-        with torch.enable_grad():
-            # Temporarily enable gradient computation for conducting the optimization
-            for _ in range(self.n_inner_iter):
-                loss = self.objective(x, y)
-                inner_optim.zero_grad()
-                loss.backward(inputs=params)
-                inner_optim.step()
-        return self
 
 
 def main():
@@ -135,18 +101,24 @@ def main():
     # We will use Adam to (meta-)optimize the initial parameters
     # to be adapted.
     net.train()
-    meta_opt = torchopt.Adam(net.parameters(), lr=1e-3)
+    fnet, params = functorch.make_functional(net)
+    meta_opt = torchopt.adam(lr=1e-3)
+    meta_opt_state = meta_opt.init(params)
 
     log = []
-    test(db, net, epoch=-1, log=log, args=args)
+    test(db, [params, fnet], epoch=-1, log=log, args=args)
     for epoch in range(10):
-        train(db, net, meta_opt, epoch, log, args)
-        test(db, net, epoch, log, args)
+        meta_opt, meta_opt_state = train(
+            db, [params, fnet], (meta_opt, meta_opt_state), epoch, log, args
+        )
+        test(db, [params, fnet], epoch, log, args)
         plot(log)
 
 
-def train(db, net, meta_opt, epoch, log, args):
+def train(db, net, meta_opt_and_state, epoch, log, args):
     n_train_iter = db.x_train.shape[0] // db.batchsz
+    params, fnet = net
+    meta_opt, meta_opt_state = meta_opt_and_state
     # Given this module we've created, rip out the parameters and buffers
     # and return a functional version of the module. `fnet` is stateless
     # and can be called with `fnet(params, buffers, args, kwargs)`
@@ -161,32 +133,40 @@ def train(db, net, meta_opt, epoch, log, args):
 
         n_inner_iter = args.inner_steps
         reg_param = args.reg_params
-
         qry_losses = []
         qry_accs = []
-        meta_opt.zero_grad()
+
+        init_params_copy = pytree.tree_map(
+            lambda t: t.clone().detach_().requires_grad_(requires_grad=t.requires_grad), params
+        )
 
         for i in range(task_num):
             # Optimize the likelihood of the support set by taking
             # gradient steps w.r.t. the model's parameters.
             # This adapts the model's meta-parameters to the task.
 
-            inner_net = InnerNet(net, n_inner_iter, reg_param)
-            optimal_inner_net = inner_net.solve(x_spt[i], y_spt[i])
-
+            optimal_params = train_imaml_inner_solver(
+                init_params_copy,
+                params,
+                (x_spt[i], y_spt[i]),
+                (fnet, n_inner_iter, reg_param),
+            )
             # The final set of adapted parameters will induce some
             # final loss and accuracy on the query dataset.
             # These will be used to update the model's meta-parameters.
-            qry_logits = optimal_inner_net(x_qry[i])
+            qry_logits = fnet(optimal_params, x_qry[i])
             qry_loss = F.cross_entropy(qry_logits, y_qry[i])
+            # Update the model's meta-parameters to optimize the query
+            # losses across all of the tasks sampled in this batch.
+            # qry_loss = qry_loss / task_num  # scale gradients
+            meta_grads = torch.autograd.grad(qry_loss / task_num, params)
+            meta_updates, meta_opt_state = meta_opt.update(meta_grads, meta_opt_state)
+            params = torchopt.apply_updates(params, meta_updates)
             qry_acc = (qry_logits.argmax(dim=1) == y_qry[i]).float().mean()
-            qry_losses.append(qry_loss)
+            qry_losses.append(qry_loss.item())
             qry_accs.append(qry_acc.item())
 
-        qry_losses = torch.mean(torch.stack(qry_losses))
-        qry_losses.backward()
-        meta_opt.step()
-        qry_losses = qry_losses.item()
+        qry_losses = np.mean(qry_losses)
         qry_accs = 100.0 * np.mean(qry_accs)
         i = epoch + float(batch_idx) / n_train_iter
         iter_time = time.time() - start_time
@@ -205,6 +185,8 @@ def train(db, net, meta_opt, epoch, log, args):
             }
         )
 
+    return (meta_opt, meta_opt_state)
+
 
 def test(db, net, epoch, log, args):
     # Crucially in our testing procedure here, we do *not* fine-tune
@@ -212,6 +194,8 @@ def test(db, net, epoch, log, args):
     # Most research papers using MAML for this task do an extra
     # stage of fine-tuning here that should be added if you are
     # adapting this code for research.
+    params, fnet = net
+    # fnet, params, buffers = functorch.make_functional_with_buffers(net)
     n_test_iter = db.x_test.shape[0] // db.batchsz
 
     qry_losses = []
@@ -221,6 +205,9 @@ def test(db, net, epoch, log, args):
     # doesn't have to be duplicated between `train` and `test`?
     n_inner_iter = args.inner_steps
     reg_param = args.reg_params
+    init_params_copy = pytree.tree_map(
+        lambda t: t.clone().detach_().requires_grad_(requires_grad=t.requires_grad), params
+    )
 
     for batch_idx in range(n_test_iter):
         x_spt, y_spt, x_qry, y_qry = db.next('test')
@@ -232,12 +219,15 @@ def test(db, net, epoch, log, args):
             # gradient steps w.r.t. the model's parameters.
             # This adapts the model's meta-parameters to the task.
 
-            inner_net = InnerNet(net, n_inner_iter, reg_param)
-            with torch.no_grad():
-                optimal_inner_net = inner_net.solve(x_spt[i], y_spt[i])
+            optimal_params = test_imaml_inner_solver(
+                init_params_copy,
+                params,
+                (x_spt[i], y_spt[i]),
+                (fnet, n_inner_iter, reg_param),
+            )
 
             # The query loss and acc induced by these parameters.
-            qry_logits = optimal_inner_net(x_qry[i])
+            qry_logits = fnet(optimal_params, x_qry[i])
             qry_loss = F.cross_entropy(qry_logits, y_qry[i])
             qry_acc = (qry_logits.argmax(dim=1) == y_qry[i]).float().mean()
             qry_losses.append(qry_loss.item())
@@ -259,6 +249,69 @@ def test(db, net, epoch, log, args):
     )
 
 
+def imaml_objective(optimal_params, init_params, data, aux):
+    x_spt, y_spt = data
+    fnet, n_inner_iter, reg_param = aux
+    y_pred = fnet(optimal_params, x_spt)
+    regularization_loss = 0
+    for p1, p2 in zip(optimal_params, init_params):
+        regularization_loss += 0.5 * reg_param * torch.sum(torch.square(p1 - p2))
+    loss = F.cross_entropy(y_pred, y_spt) + regularization_loss
+    return loss
+
+
+@torchopt.diff.implicit.custom_root(
+    functorch.grad(imaml_objective, argnums=0),
+    argnums=1,
+    has_aux=False,
+    solve=torchopt.linear_solve.solve_normal_cg(maxiter=5, atol=0),
+)
+def train_imaml_inner_solver(init_params_copy, init_params, data, aux):
+    x_spt, y_spt = data
+    fnet, n_inner_iter, reg_param = aux
+    # Initial functional optimizer based on TorchOpt
+    params = init_params_copy
+    inner_opt = torchopt.sgd(lr=1e-1)
+    inner_opt_state = inner_opt.init(params)
+    with torch.enable_grad():
+        # Temporarily enable gradient computation for conducting the optimization
+        for _ in range(n_inner_iter):
+            pred = fnet(params, x_spt)
+            loss = F.cross_entropy(pred, y_spt)  # compute loss
+            # Compute regularization loss
+            regularization_loss = 0
+            for p1, p2 in zip(params, init_params):
+                regularization_loss += 0.5 * reg_param * torch.sum(torch.square(p1 - p2))
+            final_loss = loss + regularization_loss
+            grads = torch.autograd.grad(final_loss, params)  # compute gradients
+            updates, inner_opt_state = inner_opt.update(grads, inner_opt_state)  # get updates
+            params = torchopt.apply_updates(params, updates)
+    return params
+
+
+def test_imaml_inner_solver(init_params_copy, init_params, data, aux):
+    x_spt, y_spt = data
+    fnet, n_inner_iter, reg_param = aux
+    # Initial functional optimizer based on TorchOpt
+    params = init_params_copy
+    inner_opt = torchopt.sgd(lr=1e-1)
+    inner_opt_state = inner_opt.init(params)
+    with torch.enable_grad():
+        # Temporarily enable gradient computation for conducting the optimization
+        for _ in range(n_inner_iter):
+            pred = fnet(params, x_spt)
+            loss = F.cross_entropy(pred, y_spt)  # compute loss
+            # Compute regularization loss
+            regularization_loss = 0
+            for p1, p2 in zip(params, init_params):
+                regularization_loss += 0.5 * reg_param * torch.sum(torch.square(p1 - p2))
+            final_loss = loss + regularization_loss
+            grads = torch.autograd.grad(final_loss, params)  # compute gradients
+            updates, inner_opt_state = inner_opt.update(grads, inner_opt_state)  # get updates
+            params = torchopt.apply_updates(params, updates)
+    return params
+
+
 def plot(log):
     # Generally you should pull your plotting code out of your training
     # script but we are doing it here for brevity.
@@ -272,10 +325,10 @@ def plot(log):
     ax.set_xlabel('Epoch')
     ax.set_ylabel('Accuracy')
     ax.set_ylim(80, 100)
-    ax.set_title('iMAML Omniglot')
+    ax.set_title('iMAML Omniglot (Functional)')
     ax.legend(ncol=2, loc='lower right')
     fig.tight_layout()
-    fname = 'imaml-accs.png'
+    fname = 'imaml-accs-functional.png'
     print(f'--- Plotting accuracy to {fname}')
     fig.savefig(fname)
     plt.close(fig)